Redefine Dust helpers and filters when the current tenant changes
Dust is a global resource shared by all tenants. Some helpers and filters depend on DecentCMS modules, and thus can vary in settings and state per tenant. This could cause settings to leak between tenants (example: summarizing strategy from one site applied to another). This fixes the issue by resetting those helpers and filters if the current site is different from the last that was set on the Dust object.